Last commit for the night, display loop logic MOSTLY complete

+++ src/main/java/com/pluralsight/Dealership.java
@@ -39,7 +39,7 @@ public class Dealership {
         this.phone = phone;
     }
 
-    public ArrayList<Vehicle> getVehicleByPrice(int min, int max){
+    public ArrayList<Vehicle> getVehicleByPrice(double min, double max){
         ArrayList<Vehicle> results = new ArrayList<>();
         return results;
     }
